exception from touint32 is input -1, now use toint64 and then cast to uint to get 0xffffffff, add copy operator to operator1child and check of null by index errors

+++ src/Assembler/Commands/Runtime/Asm/Command1RegisterConst.cs
@@ -97,7 +97,7 @@ namespace Yama.Assembler.Runtime
 
             if (request.WithMapper) request.Result.Add(new Command1RegisterConst(this, request.Node, assembleFormat.Result));
 
-            byte[] tmp = BitConverter.GetBytes ( Convert.ToUInt32(t.Argument1.Token.Value) );
+            byte[] tmp = BitConverter.GetBytes ( (uint)Convert.ToInt64(t.Argument1.Token.Value) );
             assembleFormat.Result.Add ( tmp[0] );
             assembleFormat.Result.Add ( tmp[1] );
             assembleFormat.Result.Add ( tmp[2] );
